refactor: 비교 페이지 유틸 함수를 entities/compare/lib/utils.ts로 이동
